requirements.txt
qdrant-client==1.9.1
redis==5.0.1
python-dotenv==1.0.0